client: use Promise.reject() over throw.
While both appear to work in the testing cases,
it appears that the gRPC client interceptors do
not handle thrown exceptions gracefully in the
callbacks. The resultant behaviour is that the
process crashes, without anyway to intercept the
error.

Using Promise.reject() instead alleviates this
problem, allowing callers to gracefully handle
errors.

Note: no tests can be written at this time to test
the behaviour, since we're currently mocking at a
level that bypasses the interceptor. The only way
to test is is to either:

  1. Run the examples against a webhook that rejects.
  2. Overhaul the existing testing framework.

(1) sufficies for now, but we may wish to revisit (2)
in the future.

GitOrigin-RevId: 5c3e459ab74a7bee8e7306f4e89fade3775a92d3