client: improve BadNonce collision handling
Previously, if a BadNonce was detected, the
retry loop would just immediately request the
account SequenceNumber and try again. The
attempting to refresh the SequenceNumber would
yield an up-to-date value.

However, it appears that the SequenceNumber is
only updated when a ledger is published to horizon.
As a result, the majority of SequenceNumber refreshes,
in the case of concurrent access, yield an outdated
value.

By optimistically predicting the next SequenceNumber
in the presence of a BadNonce, we can drastically
improve the probability that a transaction will be
accepted within the maxNonceRefreshes window.

In the case where our prediction is wrong, or we
drift too much, BadNonce will still be returned to
the caller. In this case, it is expected the caller
will retry, and the prediction state will be cleared.

GitOrigin-RevId: 45fdd0ca3745e88854db8da6fe59abdc487832d1